+++ src/main/java/com/github/idea/ginkgo/GinkgoRunConfigurationProducer.java
@@ -1,13 +1,19 @@
 package com.github.idea.ginkgo;
 
 import com.github.idea.ginkgo.scope.GinkgoScope;
+import com.github.idea.ginkgo.util.GinkgoUtil;
+import com.goide.execution.testing.GoTestFinder;
 import com.goide.psi.GoCallExpr;
+import com.goide.util.GoUtil;
 import com.intellij.execution.actions.ConfigurationContext;
 import com.intellij.execution.actions.LazyRunConfigurationProducer;
 import com.intellij.execution.configurations.ConfigurationFactory;
 import com.intellij.openapi.util.Ref;
+import com.intellij.psi.PsiDirectory;
 import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiFileSystemItem;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
@@ -23,24 +29,60 @@ public class GinkgoRunConfigurationProducer extends LazyRunConfigurationProducer
 
     @Override
     protected boolean setupConfigurationFromContext(@NotNull GinkgoRunConfiguration configuration, @NotNull ConfigurationContext context, @NotNull Ref<PsiElement> sourceElement) {
-        if (!(context.getPsiLocation().getParent().getParent() instanceof GoCallExpr)) {
+        PsiElement contextElement = getContextElement(context);
+
+        // Provides configuration for directories
+        if (contextElement instanceof PsiDirectory) {
+            PsiDirectory directory = (PsiDirectory) contextElement;
+            if (Arrays.stream(directory.getFiles()).filter(GoTestFinder::isTestFile).count() > 0) {
+                GinkgoRunConfigurationOptions options = configuration.getOptions();
+
+                GinkgoRunConfigurationOptions ginkgoRunConfigurationOptions = new GinkgoRunConfigurationOptions();
+                ginkgoRunConfigurationOptions.setGinkgoExecutable(options.getGinkgoExecutable());
+                ginkgoRunConfigurationOptions.setWorkingDir(directory.getVirtualFile().getPath());
+                ginkgoRunConfigurationOptions.setEnvData(options.getEnvData());
+                ginkgoRunConfigurationOptions.setGinkgoScope(GinkgoScope.ALL);
+                ginkgoRunConfigurationOptions.setTestNames(Arrays.asList("ginkgo", directory.getName()));
+
+                configuration.setOptions(ginkgoRunConfigurationOptions);
+                configuration.setGeneratedName();
+                return true;
+            }
             return false;
         }
 
-        List<String> specNames = getSpecNames(context);
-        GinkgoRunConfigurationOptions options = configuration.getOptions();
+        // Provides configuration for Specs and Suites
+        if ((contextElement.getParent().getParent() instanceof GoCallExpr)) {
+            GoCallExpr function = (GoCallExpr) contextElement.getParent().getParent();
+            GinkgoRunConfigurationOptions options = configuration.getOptions();
+            GinkgoRunConfigurationOptions ginkgoRunConfigurationOptions = new GinkgoRunConfigurationOptions();
+            ginkgoRunConfigurationOptions.setGinkgoExecutable(options.getGinkgoExecutable());
+            ginkgoRunConfigurationOptions.setWorkingDir(context.getPsiLocation().getContainingFile().getContainingDirectory().getVirtualFile().getPath());
+            ginkgoRunConfigurationOptions.setEnvData(options.getEnvData());
+            if (GinkgoUtil.isGinkgoFunction(function)) {
+                List<String> specNames = getSpecNames(context);
+                ginkgoRunConfigurationOptions.setGinkgoScope(GinkgoScope.FOCUS);
+                ginkgoRunConfigurationOptions.setTestNames(specNames);
+                ginkgoRunConfigurationOptions.setFocusTestExpression(String.join(" ", specNames));
+            }
 
-        GinkgoRunConfigurationOptions ginkgoRunConfigurationOptions = new GinkgoRunConfigurationOptions();
-        ginkgoRunConfigurationOptions.setGinkgoExecutable(options.getGinkgoExecutable());
-        ginkgoRunConfigurationOptions.setWorkingDir(context.getPsiLocation().getContainingFile().getContainingDirectory().getVirtualFile().getPath());
-        ginkgoRunConfigurationOptions.setEnvData(options.getEnvData());
-        ginkgoRunConfigurationOptions.setGinkgoScope(GinkgoScope.FOCUS);
-        ginkgoRunConfigurationOptions.setTestNames(specNames);
-        ginkgoRunConfigurationOptions.setFocusTestExpression(String.join(" ", specNames));
+            if (GinkgoUtil.isGinkgoSuite(function)) {
+                String suiteName = getSuiteName(function);
+                ginkgoRunConfigurationOptions.setGinkgoScope(GinkgoScope.ALL);
+                ginkgoRunConfigurationOptions.setTestNames(Arrays.asList("ginkgo", suiteName));
+            }
 
-        configuration.setOptions(ginkgoRunConfigurationOptions);
-        configuration.setGeneratedName();
-        return true;
+            configuration.setOptions(ginkgoRunConfigurationOptions);
+            configuration.setGeneratedName();
+            return true;
+        }
+
+        return false;
+    }
+
+    @NotNull
+    private String getSuiteName(GoCallExpr function) {
+        return function.getArgumentList().getExpressionList().get(1).getText().replace("\"", "");
     }
 
     @Override
@@ -77,5 +119,20 @@ public class GinkgoRunConfigurationProducer extends LazyRunConfigurationProducer
     @Override
     public ConfigurationFactory getConfigurationFactory() {
         return ginkgoConfigurationFactory;
+    }
+
+    @Nullable
+    protected PsiElement getContextElement(@Nullable ConfigurationContext context) {
+        if (context == null) {
+            return null;
+        } else {
+            PsiElement psiElement = context.getPsiLocation();
+            if (psiElement != null && psiElement.isValid()) {
+                PsiFileSystemItem psiFile = psiElement instanceof PsiFileSystemItem ? (PsiFileSystemItem) psiElement : psiElement.getContainingFile();
+                return !GoUtil.isInProject(psiElement.getProject(), psiFile != null ? psiFile.getVirtualFile() : null) ? null : psiElement;
+            } else {
+                return null;
+            }
+        }
     }
 }