+++ src/main/java/com/github/idea/ginkgo/util/GinkgoUtil.java
@@ -2,6 +2,7 @@ package com.github.idea.ginkgo.util;
 
 import com.github.idea.ginkgo.GinkgoPendingSpecType;
 import com.github.idea.ginkgo.GinkgoSpecType;
+import com.goide.psi.GoCallExpr;
 
 import java.util.Arrays;
 
@@ -11,11 +12,17 @@ public class GinkgoUtil {
         //Util class should not be instantiated.
     }
 
-    public static boolean isGinkgoFunction(String name) {
+    public static boolean isGinkgoSuite(GoCallExpr function) {
+        return function.getExpression().getText().equals("RunSpecs");
+    }
+
+    public static boolean isGinkgoFunction(GoCallExpr function) {
+        String name = function.getExpression().getText();
         return Arrays.stream(GinkgoSpecType.class.getEnumConstants()).anyMatch(e -> e.specType().equals(name));
     }
 
-    public static boolean isGinkgoPendingFunction(String name) {
+    public static boolean isGinkgoPendingFunction(GoCallExpr function) {
+        String name = function.getExpression().getText();
         return Arrays.stream(GinkgoPendingSpecType.class.getEnumConstants()).anyMatch(e -> e.specType().equals(name));
     }